file structure updates
