+++ x-pack/platform/plugins/shared/onechat/public/application/components/conversations/embedded/rounds_layout.tsx
@@ -1,0 +1,14 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React from 'react';
+
+export const EmbeddedConversationRounds: React.FC<{
+  children: React.ReactNode;
+}> = ({ children }) => {
+  return <div>{children}</div>;
+};
