+++ x-pack/platform/plugins/shared/onechat/public/application/pages/conversations.tsx
@@ -8,7 +8,7 @@
 import React from 'react';
 import { useBreadcrumb } from '../hooks/use_breadcrumbs';
 import { appPaths } from '../utils/app_paths';
-import { OnechatConversationsView } from '../components/conversations/conversations_view';
+import { OnechatConversationsView } from '../components/conversations/full_screen/conversations_view';
 import { labels } from '../utils/i18n';
 
 export const OnechatConversationsPage: React.FC = () => {
