+++ x-pack/platform/plugins/shared/onechat/public/embeddable/embeddable_conversation.tsx
@@ -13,7 +13,7 @@ import { css } from '@emotion/react';
 import type { EmbeddableConversationInternalProps } from './types';
 import { EmbeddableConversationsProvider } from '../application/context/conversation/embeddable_conversations_provider';
 import { Conversation } from '../application/components/conversations/conversation';
-import { EmbeddedConversationHeader } from '../application/components/conversations/embedded_conversation_header';
+import { EmbeddedConversationHeader } from '../application/components/conversations/embedded/header';
 
 export const EmbeddableConversationInternal: React.FC<EmbeddableConversationInternalProps> = (
   props
